pyproject.toml
[tool.poetry]
name = "dnssec-keysize"
version = "0.1.0"
description = ""
authors = ["Jakob Schlyter <jakob@kirei.se>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
dnspython = "^2.3.0"
cryptography = "^39.0.1"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
